+++ package.json
@@ -46,6 +46,7 @@
   "homepage": "https://github.com/yomariano/masterstake#readme",
   "dependencies": {
     "fastify": "^1.1.1",
+    "fastify-accepts": "^0.5.0",
     "fastify-multipart": "^0.4.1",
     "fastify-plugin": "^0.2.2",
     "mongoose": "^5.0.10",
@@ -55,6 +56,7 @@
     "husky": "^0.14.3",
     "lint-staged": "^7.0.0",
     "nodemon": "^1.17.2",
+    "prettier": "^1.11.1",
     "tap": "^11.1.2"
   }
 }
